Rename services/finp2p-contract to services/onchain; unify into OnChainTokenService
The folder is named after the on-chain execution model, not the provider-type
string (which stays 'finp2p-contract' in config).

- CommonServiceImpl + TokenServiceImpl + EscrowServiceImpl collapse into a
  single OnChainTokenService implementing TokenService, EscrowService,
  CommonService and HealthService — one contract client, one credentials
  cache, one instance wired for the token/escrow/common registrations
  (drops the `tokenService as any` fallback).
- model.ts (BusinessContract; the ExecutionContext type was dead) and
  validator.ts (validateRequest) fold into helpers.ts — the EIP-712
  business-details types, extraction and validation now live together.
- The ExecDetailsStore interface moves next to its in-memory implementation
  in exec-details-store.ts.
