--- src/services/finp2p-contract/common.ts
@@ -1,66 +1,0 @@
-import {
-  CommonService, HealthService, OperationStatus,
-  ProofProvider, PluginManager,
-  ReceiptOperation, ExecutionContext,
-} from "@owneraio/finp2p-nodejs-skeleton-adapter";
-import { FinP2PContract } from "@owneraio/finp2p-ethereum-orchestrator";
-import { FinP2PClient } from "@owneraio/finp2p-client";
-import { mapReceiptOperation } from "./mapping";
-
-
-export interface ExecDetailsStore {
-  addExecutionContext(txHash: string, executionPlanId: string, instructionSequenceNumber: number): void;
-
-  getExecutionContext(txHash: string): ExecutionContext;
-}
-
-export class CommonServiceImpl implements CommonService, HealthService {
-
-  finP2PContract: FinP2PContract;
-  finP2PClient: FinP2PClient | undefined;
-  execDetailsStore: ExecDetailsStore | undefined;
-  proofProvider: ProofProvider | undefined;
-  pluginManager: PluginManager | undefined;
-
-  private readonly registeredCredentials = new Set<string>();
-
-  constructor(
-    finP2PContract: FinP2PContract,
-    finP2PClient: FinP2PClient | undefined,
-    execDetailsStore: ExecDetailsStore | undefined,
-    proofProvider: ProofProvider | undefined,
-    pluginManager: PluginManager | undefined
-  ) {
-    this.finP2PContract = finP2PContract;
-    this.finP2PClient = finP2PClient;
-    this.execDetailsStore = execDetailsStore;
-    this.proofProvider = proofProvider;
-    this.pluginManager = pluginManager;
-  }
-
-  protected async ensureCredential(finId: string): Promise<void> {
-    if (this.registeredCredentials.has(finId)) return;
-    await this.finP2PContract.getCredentialAddress(finId);
-    this.registeredCredentials.add(finId);
-  }
-
-  public async readiness() {
-    await this.finP2PContract.provider.getNetwork();
-  }
-
-  public async liveness() {
-    await this.finP2PContract.provider.getBlockNumber();
-  }
-
-  public async getReceipt(id: string): Promise<ReceiptOperation> {
-    return mapReceiptOperation(await this.finP2PContract.getReceipt(id), undefined, this.execDetailsStore?.getExecutionContext(id));
-  }
-
-  public async operationStatus(cid: string): Promise<OperationStatus> {
-    const op = await this.finP2PContract.getOperationStatus(cid);
-    if (op.operation === 'receipt') return mapReceiptOperation(op, undefined, this.execDetailsStore?.getExecutionContext(cid));
-    return op as any;
-  }
-
-
-}
