--- src/services/finp2p-contract/index.ts
@@ -1,6 +1,0 @@
-export * from './common'
-export * from './escrow'
-export * from './exec-details-store'
-export * from './mapping'
-export * from './model'
-export * from './tokens'
