--- src/services/finp2p-contract/model.ts
@@ -1,15 +1,0 @@
-import { EIP712LoanTerms, OperationParams, Term } from "@owneraio/finp2p-ethereum-orchestrator";
-
-export type BusinessContract = {
-  buyerFinId: string,
-  sellerFinId: string,
-  asset: Term,
-  settlement: Term,
-  loan: EIP712LoanTerms,
-  params: OperationParams
-};
-
-export type ExecutionContext = {
-  executionPlanId: string
-  instructionSequenceNumber: number
-}
